refactor(frontend): move `<AnchorLink>` inside `links.tsx`
